refactor: 폴더 구조 변경
이가은님 폴더 구조 참고
